Move completer into index file, require readline inside listener
